+++ src/modules/admin/widgets/grids/EntryGridView.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace davidhirtz\yii2\cms\tenant\modules\admin\widgets\grids;
+namespace Hirtz\Cms\tenant\modules\admin\widgets\grids;
 
-use davidhirtz\yii2\cms\tenant\modules\admin\widgets\grids\traits\EntryTenantGridViewTrait;
+use Hirtz\Cms\tenant\modules\admin\widgets\grids\traits\EntryTenantGridViewTrait;
 
-class EntryGridView extends \davidhirtz\yii2\cms\modules\admin\widgets\grids\EntryGridView
+class EntryGridView extends \Hirtz\Cms\modules\admin\widgets\grids\EntryGridView
 {
     use EntryTenantGridViewTrait;