+++ src/modules/admin/widgets/grids/SectionParentEntryGridView.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace davidhirtz\yii2\cms\tenant\modules\admin\widgets\grids;
+namespace Hirtz\Cms\tenant\modules\admin\widgets\grids;
 
-use davidhirtz\yii2\cms\tenant\modules\admin\widgets\grids\traits\EntryTenantGridViewTrait;
+use Hirtz\Cms\tenant\modules\admin\widgets\grids\traits\EntryTenantGridViewTrait;
 
-class SectionParentEntryGridView extends \davidhirtz\yii2\cms\modules\admin\widgets\grids\SectionParentEntryGridView
+class SectionParentEntryGridView extends \Hirtz\Cms\modules\admin\widgets\grids\SectionParentEntryGridView
 {
     use EntryTenantGridViewTrait;